requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
astroid==2.4.2
azure-core==1.9.0
azure-storage-blob==12.6.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
idna==2.10
isodate==0.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.19
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
bs4
